+++ src/ts/ui/initUI.ts
@@ -99,8 +99,8 @@ export const setPadding = (vditor: IVditor) => {
         outlineWidth = outlienElement.offsetWidth;
     }
 
-    vditor.toolbar.element.style.paddingLeft = parseInt(vditor[vditor.currentMode].element.style.paddingLeft, 10)
-        + outlineWidth + "px";
+    vditor.toolbar.element.style.paddingLeft =
+        parseInt(vditor[vditor.currentMode].element.style.paddingLeft || "0", 10) + outlineWidth + "px";
 };
 
 export const setTypewriterPosition = (vditor: IVditor) => {